fix: 読み込みファイルに改行があった場合などの対応
