+++ rt_files/ng/line_with_only_spaces.rt
@@ -1,0 +1,5 @@
+A 0.2 255,255,255
+C 0,10,0 0,-1,0 60
+L 200,200,200 0.8 255,255,255
+   
+sp 0,0,0 4 255,255,255
